ImageProc module: move the src of lambda functions into the src folder
