+++ ImageProcessing/src/lambda-functions/copy-s3-object/requirements.txt
