+++ ImageProcessing/src/lambda-functions/face-detection/index.js
