+++ ImageProcessing/src/lambda-functions/face-search/index.js
