+++ ImageProcessing/src/lambda-functions/face-search/test-input.json
