+++ ImageProcessing/src/lambda-functions/index-face/index.js
