+++ ImageProcessing/src/lambda-functions/mock-notification/index.js
