+++ ImageProcessing/src/lambda-functions/mock-notification/test-input.json
