+++ ImageProcessing/src/lambda-functions/persist-metadata/index.js
