+++ ImageProcessing/src/lambda-functions/persist-metadata/test-input.json
