+++ ImageProcessing/src/lambda-functions/thumbnail/index.js
