+++ ImageProcessing/src/lambda-functions/thumbnail/package.json
@@ -6,6 +6,7 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
+  "author": "Angela Wang",
   "license": "Apache 2.0",
   "dependencies": {
     "gm": "^1.23.0",
