+++ ImageProcessing/src/lambda-functions/thumbnail/test.json
